pyproject.toml
[project]
name = "temporal-streaming"
version = "0.1.0"
description = "Temporal streaming example with Redis"
requires-python = ">=3.10"

dependencies = [
    "temporalio",
    "redis",
    "anthropic",
    "python-dotenv"
]
